requirements.txt
tensorflow==2.15.0
numpy==1.26.3
matplotlib==3.8.2 
tqdm==4.66.1
opencv-python==4.9.0.80
matplotlib==3.8.2 
smbus==1.1.post2 
pyserial==3.5 